Refactored Component to be functional and renamed Piwik to Matomo
